+++ inc/lilyNotation/index.ts
@@ -2,6 +2,7 @@
 import {Note, Notation, assignNotationEventsIds} from "./notation";
 import {fuzzyMatchNotations} from "./fuzzyMatch";
 import {matchWithMIDI, matchWithExactMIDI} from "./matcher";
+import Scheduler from "./scheduler";
 import * as measureLayout from "../measureLayout";
 import {LayoutType} from "../measureLayout";
 
@@ -18,6 +19,7 @@ export {
 	assignNotationEventsIds,
 	matchWithMIDI,
 	matchWithExactMIDI,
+	Scheduler,
 	LayoutType,
 	MLayoutClasses,
 };