+++ changelog.md
@@ -16,7 +16,9 @@
 
 *	2021.1.15
 
-	*	To fix 'pure rest measure' scheduler issue, replace 'r' with pitched rest (\rest modified chord).
+	+	scoreBundle.ts: added scheduler loading.
+
+	*	To workaround 'pure rest measure' scheduler issue, replace 'r' with pitched rest (\rest modified chord).
 
 	+	sheetDocument.ts: enhanced alignTokensWithNotation by note tokens (include rests).